Interpret UserComment character code
Reads first 8 bytes for character encoding recognition to comply with
EXIF standard.

It's not necessary to read 10 bytes at the beginning. Reading these two
extra bytes could lead to missing bytes and the end. Switching the byte
order for UNICODE to big-endian solves the issue with 8 and 10 bytes
reads.

Only for ASCII NULL termination is done, "since the type is not ASCII,
NULL termination is not necessary".