IJpegSegmentMetadataReader returns enumerable
This change simplifies the code a little and improves performance, primarily by removing list allocations.

Previously this method would return a materialised collection. This is not necessary, and an enumeration of directories is sufficient.

This is an API change, but is unlikely to break many users as this interface is mostly used internally.

It also replaces runtime casts (for .NET 3.5, where interface covariance does not exist) with compile time casts.