Work around invalid XMP XML
Software that re-writes XMP can do so without changing the size of the XMP segment. If it overwrites the data with a shorter payload, there can be junk at the end which causes the XML parsing to fail.

This change adds a scan for known end-of-message markers, to avoid this problem.